Add handlers, fix tests
Add counterattack and body_part_inject_self handlers.

Fix WORLD4 body part name: head -> head1.
'head' is a body part type but an object with name 'head' will cause
the body_part property to be changed to the object's pid.

I ran into a horrible bug where a handler was returning undefined from a
succeed call and poisoning the props of an object. The handler was
completely unrelated to the test that was failing. What a nightmare.
I ended up adding a ton of ct:pal statements and left one in the logger
where it prints a message if a process has 'undefined' for its
properties.

Did a whole bunch of changes to item movement:
- kick off a "Hey Item, tell me your valid body parts" message;
- Have body parts identify as 'first_available_body_part'
- Have body parts identify as 'current_body_part'
- Have items respond to item_body_parts with their body parts
  (Instead of synchronously getting the item's body parts from
   through erlmud_object:get(...))

Add the target to attack properties on characters so that the
counterattack handler can tell if the current target is the same as the
one that most recently attacked ... not even sure if that is required
now. ... at least not for the "pitbull" case of "keep attacking whatever
you're currently attacking.

Have characters inject the room on {Char, drop, Item} and {Char, get,
Item} because characters can _only_ be owned by rooms. The character has
a room property so it's easier for the character to lookup it's room
than the room to lookup if it has the character. This shortcut doesn't
work for items because items can be owned by rooms, characters,
body_parts, etc. (Starbucks is closing and I can't remember all the
details but I _thought_ I could do this in other cases).

Add some checks so that when I have binary names in place of an item or
a body part in a {move, Item, from, Char, to, BodyPart} that the
character doesn't subscribe to that: it only subs to movement messages
where all the names have been converted to pids.

Oh, crap, I need to watch for cases where item and body part names don't
match.

Fixed a few tests.

Added a separate test for auto-wielding in the only available spot.