Add a broadcast attempt result
Only body parts know which character they and all their children
processes are attached to. Children of the body parts don't recognize
the character they are attached to when attack events come by. We need a
way to tell processes under a character which character they are under.

The only things we can add to the character so far in-game are items.
When we add the item to the player we can send a message to the item
telling it to add the character as a {character, Character} property.
One way to do this would be to kick off an attempt and every process can
add themselves to the attempt forming a process tree: processes could see
what character they live under. However, to make things more efficient
I've made a broadcast result to attempts that causes the specified
message to be broadcast to all of a process's attached processes except
their owner and character. Processes that get one of those messages from
their parent can re-broadcast the message to their attached processes.
This is not a single message that reaches each child by traversing the
graph, this is a separate message for each process. The broadcast result
type only handles sending a different message to each attached process
and doesn't handle channeling or blocking them.
The erlmud_handler_set_character module, which I've added to every
object, blocks the message from going past non-children and re-broadcasts
it from children to grandchildren.

So an item will send a message to all of its attached processes except
the owner and (new) character property. Immediate children of the
process will recognize the message and re-broadcast it with themselves
as the owner. Processes that don't recogize the "Owner" as their owner
will fail the message. If one message was sent out to traverse the graph
then a fail would stop the message before every child process had seen
it. This change causes a burst of short lived messages that are failed
(and stop propagating) as soon as they reach a process that isn't a
direct child of the owner. A wave of messages will flood down the
character process tree and be blocked from travelling "up". Since every
attached process, and therefore every child process, gets their own copy
of the message, failing the messages doesn't prevent any process from
seeing it.

This is used to broadcast a {set_character, Owner, Character} to all the
children of a character so that they can recognize the character during
attacks. For instance, if the player is wielding a sword in their hand
and the sword has a spell on it, then the sword and the spell will know
that attacks involving their character parent involve them.

Added {character, X} properties to a bunch of test objects to
make the existing tests work.

Added a test for broadcasting {character, _, _}.