Finish & Fix tests for attack modifiers
Remove all the "Object types" from the test worlds. The only type
information is in the property keys in object properties. For example, a
room will have a property with a key 'room'. Although the attached room
doesn't know that it's a room, the object pointing to it assumes that
its. Objects are differentiated not by their intrinsic types but by
their event handlers, i.e. their behaviour.

Add some attributes to various test characters and objects so that
attacks can hit and do damage. Players no longer have a default ability
to hit.

Remove the body part attack handler. I copied this verbatim from another
handler and didn't need it.

Add item and attribute attack handlers that listen for calc_hit and
calc_damage events. The handlers will add or subtract hit and damage
modifiers depending on if they are the target or the source of the
attack.