Test character property unset on item drop
When a character picks up an item made up of mulitple processes each
of the processes should have their character property set. Likewise,
when the character drops the item the character property should be
unset.

This should happen whenever the character loses possession of the item,
not just when the character drops the item.