Set item hierarchy body_part on wield/wear
Complex items will have attributes, spells and stats that have different
effects while wielded (or worn) rather than sitting in inventory.

In order for each process in an item hierarchy to recognize that they
are wielded (or worn) when an attack is in progress they need to know
that they have a body_part.

For example, a regular helmet that is in inventory won't affect hit rolls.
This doesn't preclude items having an effect while in inventory, it just
allows us to differentiate. Items in general inventory will have the
character as their owner and can pattern match on messages that have
their owner.