Combine child property setters, add top_item
In order to particpate in attacks all items in an item hierarchy will
need to know what character owns them, what body part wears/wields them
(if any) and what the top-level item is. If multiple attacks go by each
sub-item will need to affect only the attaks involving their top-level
item.

If a Bob attacks Alice then Bob's gun needs to be able to recognize "Bob
attacks Alice" even if the gun's owner is Bob's left hand.

If Bob attacks Alice then Bob's gun also needs to know if Bob is
actually holding the gun, or if Bob is wearing it for a hat. Bob's gun
might then kick off "Bob attacks Alice with Gun".

If Bob attacks Alice and Bob is holding a gun with a silencer, a clip
full of explosive rounds, a laser site, a custom grip and a flashlight,
then all of those sub-items need to be able to recognize that Bob is
their owner, they are all being wielded and that they are part of the
current attack.
For example:
"Bob attacks Alice with pistol."
    The explosive rounds in Bob's gun are going to particpate in this.
vs.
"Bob attacks Alice with wet noodle."
    Every part of Bob's gun is going to ignore this.