Update tick at reservation, allocate resources
Instead of sending an update_tick event to update the properties of the
resource I'm directly editing the properties right in the reservation
handler. I don't think we'll need to block the update of a tick. (i.e.
anything we send as an attempt can be blocked.)

Add allocation code: once the resource accumulates it checks if there's
enough resource to allocate to the next process in line.