WIP: more work switching over to attack resources
Instead of an attack process that coordinates attacks for all items,
spells, body parts, etc. each process that can attack will handle its
own attack.

This means each item and sub-item need to know what their top level
"attacker" process is, since it will be the character triggering the
attack in response to a character command or some kind of character
based event (e.g. auto-retaliate).

I'm taking out the code where the attack process is a separate process
and converting it to be a handler that is attached to a child process of
the character.