WIP: convert objects to attack
Resources will be shared by every process belonging to a character, so
the resource will only listen for reservations from processes that share
the same character. In the case of resources the Character will probably
be the Owner, but this is more explicit.

Any attacking process (spell, item, body part, etc.) needs to check that
the attacker is its Character. For example, if I'm a spell I'm only
going to listen for attacks from the Character that owns me.