Add #parents{}
Until I switch to maps and I can bind on map values I'm going to use a
parents{} record so I can bind on record values. This way I can match
not only on the owner but also the character, body part and top item.
I need to bind on these things because whether or not processes are
interesting in attacking will depend on whether the character is
involved, whether a sub-item's parent item is actually participating in
the attack whether the attacking top level item is wielded or worn on
a body part.

Remember that these are from the properties for an item: e.g. an item
process representing a gun may belong to a character, and may be wielded
in a hand. Also, a silencer may have those in addition to having a gun
as a top-level item.