Adding comments
Even though I'm a proponent of no comments I'm struggling to figure out
the logic based on binding on incoming events. Each handler module
matches on incoming events using pattern matching but I haven't taken the
time to figure out a way to make this expressive. For now, especially
while I transition to multi-attacks with resources, I'm going to put in
comments. This way when I pick at away at this once in a while I can
quickly get back up to speed. I really don't like comments though: they
aren't checked for accuracy by the compiler and they can lead to lazy
programming where the code is unreadable and comments are added as a
bandaid: if the code changes it's not easy to reconcile the code against
the comment.