Item handlers check if wielded and active
When a general attack is kicked off each process with a generic attack
handler will resend the attack with itself as the attack "vector". Item
processes will then check if the item is wielded on an appropriate body
part and activated. For instance, a character could have a free hand (or
a thousand free hands, or a free tentacle, or a free robotic appendage,
Erlmud shouldn't care) but not want to attack with a bare hand. Also, a
sword may require that it is wielded by some kind of hand-type
appendage, like a tentacle, robotic clamp, hand, etc.