Move attack to spell_attack
I'm not using a generic attack handler anymore until I find a nice way
to delegate some of the events to a process-specific handler: e.g. have
sub-item processes being able to handle some of the specifics of an
attack but have a generic handler take care of the rest.