More multi_attack
Gutted the erlmud_handler_char_attack as there's nothing left for it to
do. All the attacking is now handled by individual attack vectors:
spells, weapons, ... whatever else.

Swtiching to use a #parents{} record that holds the Owner, TopItem,
Character and BodyPart. This allows me to match on various parent items
much easier. For example, a sub-item can easily check if its top item is
involved in an event. For sub-items that are one level deep that would
just be "Owner", but for more than one level deep the sub-sub-...-items
need a reference to the top-most item.

Switching to verbose attack handler events for clarity. Instead of
having an #attack{} record that tracks the target, source, weapon, calc
phase, etc. I'll just specify each event manually:
{Character, calc, Hit, on, Target, with, AttackVecor}
I find this much easier to follow even if more verbose.

Counterattack will be handled by individual attack vectors since those
are the things that are actually attacking. It seems weird to have a
weapon or spell in charge of attacking with itself, but I don't see the
need yet for separate attack processes. The attack vectors know what
resources they require to attack. ... Or at least when the world is put
together that will be the case. It makes sense: a sword would have
require melee resources, a spell would require spell casting resources,
etc.