Link non-{K,V} Pids with ?PID() macro
When the test worlds start all the objects are started and their names
are mapped to their PIDs. After that each {K,V} property of each object
is checked to see if V matches the name of a process; if it does then V
is converted from that name (an atom) to the PID that matches that name.

This adds a new way to match processes in properties. It is in no way a
general solution and therefore it is a hack. However, it is bed time and
I want to check this in. :)

By specifying a {K, {V1, V2}} value and wrapping V1 in {pid, V1}, V1 can
also be replaced with the PID who's name matches V1. I've added a handy
dandy macro for this: ?PID(V1). This simply wraps V1 in {pid, V1}.
erlmud_object will pick up on this--very--specific pattern and reaplce
{K, {{pid, V1}, V2}} with {K, {PidOfV1, V2}}. ?PID should stand out more
than just {pid, NameOfObject}.

I "needed" this because I want to store a {body_part, {PID, Type}}
property on children of body parts. This way I can check if Type is
valid because Pid will be a valid Pid. It's not really necessary but I
decided I liked it this way. It's my party and I can kludge if I want
to.