Fixing up item attacks, ... not working yet.
This fixes some stuff that was wrong in the item attack handler like
is_wielded.

Next I need to add resources to objects in the test worlds starting with
world 3 which is used for the player attack test; the player attack test
is one of the main focuses of the multi_attack branch.

I think I might add another erlmud_object:set_pid/2 possibility so that
I can have lists of resources in a property allong with their type ...
or something. I dunno.