Fix player_attack test
The attacking item doesn't get a second chance to modify the "calc hit"
or "calc damage" events so it needs to set the initial values when the
attempt is kicked off.

I gave the player's fist in World 3 a default "calc hit" value. All
objects that initiate "calc hit" can reasonably default the intial value
to 1. It's fair to assume that if nothing prevents the hit that it will
connect.