Fix tests
Change the attack_wait test to a resource_wait test: there is no more
attack process: no waiting is controlled by how long it takes the
resource to refill itself. In the player_resource_wait the stamina
resource starts out with 5, which is enough for one attack, but the
resource tick_time is so long that the player never gets enough stamina
to attack again.

Fix world 10 where the room's player reference was incorrect.