WIP: going back to counterattack
I started ditching the counterattack events because an object can only
handle an event with one handler: if the counterattack handler handles
an event that succeeds, then the item attack handler won't see the
successful event. That was making things annoying.

However, I need to differentiate between a player command to attack a
target and a counterattack because the target attacked us.

So, the character will kick off a counterattack when the
erlmud_handler_char_attack handler senses an attack. the item attack
handlers will listen for a counterattack event and only subscribe if
they are not already engaged. When the player issues an attack command
then all of the item attack handlers will respond regardless.

This does bring up an interesting situation where a character is
attacking with only some of it's weapons. Once the target counterattacks
the character the character will then counterattack the targets
counterattack.At this point any of the character's items that were not
engaged in the attack will start to attack. If the player decided to
only attack with spells (maybe other attacks, like explosives, are too
dangerous, or costly) then once the attack -> counterattack ->
counterattack cycle happens all weapons will be engaged.

e.g.

- Character attacks Enemy with cheap weapon
- Enemy sees attack, counterattacks Character
- Character sees attack from Enemey
  (because the Enemey counterattack kicked off the enemy attack)
  and kicks off its own counterattack.
- The cheap weapon is already attacking the Target and so ignores
  the counterattack event.
- An expensive weapon is _not_ attacking, sees the counterattack,
  and starts attacking the target.

Solutions:
- Only counterattack with items that were originally attacked with?
  At this point the character could store that it is attacking and just
  not counterattack.
- Have a default counterattack set of items that will counterattack.
  That's going to get confusing for the player:
  "Why am I attacking with this expensive weapon when I specified a
   cheap weapon?" -> "Oh, you didn't specify not to _counterattack_
   with the expensive weapon."