Fix attack /w modifiers test
Items that can attack now have an 'is_attack' property and items that
can defend have an 'is_defence' property. The 'is_active' property
didn't model weapons that were active for defence but not attack (like a
shield).

The character now kicks off counter attacks and anything that can attack
will start attacking if it isn't already.

I fixed test world 8 so that the giant had stamina and the giant's club
required stamina. I'm probably going to regret tieing stamina
requirements directly to weapons, instead of the relationship between a
character and a weapon, but I can always change that using an attribute
of the character and an attribute of the weapon: e.g. strength of
character and weight of weapon.

The test passes: the player attacks the giant, the giant counter
attacks, the player does damage but the giant cannot. The player lives
without a scratch and the giant dies.