Fix one-sided (and then counterattack) test
To fix the one-sided test I gave the player character a dexterity that
added 99 to the to_hit defence. This meant that the counterattack test
had to be fixed because the zombie couldn't land the initial hit that
kicks off the player counterattack. (I suppose it would be valid to have
just a failed attack kick off a counterattack, but I'll leave the test
as is for now).