Fixed a few tests: body_part returns part type
The body part property on any object will return both the PID of the
body part that owns the object as well as the body part type. I don't
know why this is required anymore, but maybe something to do with
needing to know if this body type is listed as one of the "wield" body
parts or if it's just hung on the body somewhere (e.g. handing a shield
on your back.)