Log the Pid of object with their properties
It seems that the properties that are written to the HTMl log file are
getting out of order with the events that are getting logged. This lists
the pid of the object that the properties belong to so that it's easy to
tell if they don't match the event target.