Use refs to remove correct properties
A body part property should only be removed from an item until a new
body part property is set. That is, if an item goes from body parts A to
B to A, then the body part property should only be removed if it's from
the _first_ time that A was the body part, not the second time.

Lots of stuff changes to keep track of the ref associated with the item
property on the body part (so it can remove itself) and the ref
associated with the body part on the item.