Expect body part refs
Every time a body part has a new item descendent it creates a ref to
mark the current relationship and passes it on to the item. The item
listens for a "clear property" event to clear the body part with that
ref. If the clear property event comes out of order (i.e. when a new
body part now owns the item) then the command is ignored. This handles
the case where the clear property event (to clear the body part
property) arrives out of order.

Some description tests are still broken.