Store starting room with player
Instead of trying to have a global start room store the start room with
the player. When new players are created they will need a start room.
When existing players connect they should know what room they are in:
either the room they were in last or somewhere else based on game rules
(e.g. if you disconnect in an arena you reconnect outside the arena)

This means the connection gear know longer knows are cares what room the
player is in.