Only create filter checkboxes once
Apparently (since I don't really remember) when the tests would crash
the HTML log wouldn't have correct (enough, since browsers are lenient)
HTML to render properly. I added the checkbox code in two places: once
after the main log loop was complete and once in the static "footer"
file. If the log _did_ complete correctly then the code in both places
would create the checkboxes. Not only was this redundant but somehow
(and I didn't take the time to figure out why) it was also hiding
errors: once I set the checkboxes to only be created once I was seeing a
critical error. It was quite frustrating to look at the HTML log and see
the log lines end abruptly with no clue as to what went wrong.