Differentiate body part refs
Each body part will create a new ref every time it becomes an ancestor
of an item (e.g. parent, grand-parent, etc.). If the refs for multiple
body parts (or multiple items on the same body part, or even the same
item on the same body part multiple times) are bound to the same
variable then the binding will fail and break the test. In fact, we
could even test that the body part - item relationship refs are
different.