Flatten descriptions
I'm not _exactly_ sure where I was going with the deep/shallow
description markers but instead of relying on flags I've switched to
triggering on whether the process or its parent is being described. The
general idea is that if a process is looked at it describes itself and
then fires off a message for its children to describe themselves.
However, it seems a bit broken in that attributes, if looked at
directly, will fire off the exact same message as they responded to,
creating an endless loop. It seems like characters kick off the new
event and subscribe to it when they send it, rather than having an
attempt() clause that handles it: that is, a character will only run the
succeed() clause for describing itself when it itself has sent the
message in an attempt clause: there's no raw attempt clause that will
respond to it.

I suspect some of this is working by coincidence or because the tests
are lacking but I'm checking it in anyway since the current test suite
passes.

The _look handlers could using some DRYing out as they all use some very
similar deep/shallow (i.e. describe myself vs. describe my parent)
logic.