look at targets, describe children
By which I mean the {Source, look, Self} will cause the target to send
it's description and then kick off a describe command for it's children.
The target itself will ignore {_, describe, Self, _, _} commands.

However, if the item is a child and gets a describe-owner command then
it will describe itself in the context of the owner.