Only describe player on success
Previously I was describing the player during the event attempt rather
than when the event actually succeeded.

This follows the item look gear in that when a character is looked at
directly it will print its description and then kick off a describe
command for its children to trigger on. If the item is a child then it
will get describe commands when its parent/owner is looked at.

I don't know if I've noted this anywhere but this means descriptions
will only occur on two levels: the parent and the child. Previously I
think I had ideas of doing a deeper look at an item: e.g. looking at a
room will not only show you what characters are in the room, but what
items and body parts belong to those characters. I may make a new
command in the future (e.g. "examine") that does that.