Remove TODO
It does make sense that a single item will be returned because I'm using
a proplists:get_value(key, List) call which will return the first item
as a scalar value (i.e. not a list) that matches the key; it doesn't
return all the items that have that key.

1 = proplists:get_value(a, [{a, 1}, {a, 2}]).