Remove ct:pal calls, make port configurable, ...
... change the run script (r) so that it changes the event logger
directory to the log directory instead of the current working directory,
which is _rel/erlmud

I had some ct:pal calls in the code because I've only ever run the mud
through tests.

Make the port configurable in the sys.config file in case the 8080 port
is already in use.