Assume items are moved to rooms if not 'via'
Unless a move attempt has a 'via' property (i.e. an exit) then it's
assumed that the move is for an item.

I had some code that expected a Type to go along with the Object
being moved. I don't think I need it.

The erlmud_handler_room_inject_self handler was injecting itself in
get/drop attempts except the character object already does that.

erlmud_handler_room_move.erl doesn't do anything with item moves, only
character moves, so it doesn't need to catch a successful item move. It
was only logging the move, but it's confusing to have it logging a move
it had nothing to do with.