Create listing of all messages in all handlers
I need a way to wrangle all of the messages that are generated in the
MUD. For now just being able to see them in a list with their handler
file names will be a help.

Eventually I might look at finding a way to use a graph DB, or
heuristics, message keys (maybe hashes?) or macros to track what
messages are used.