WIP: add character to room upon connection
The character will always know what room it's in, but the room will only
know about the character once the character connects. If the room
doesn't know about the player then it won't be able to send any events
to the player.

I tried storing the name of the room that the player should start in,
but if I start the event with the player the message has no way to get
to the room. For now when the initial world is loaded each player will
already know what room it's in.

Eventually I'll need to handle connecting new players to a starting
room. Whatever process handles that should have a permanent link to the
starting room(s) of the MUD.