Fix look tests
The erlmud_handler_char_enter_world handler was listening for the
final {Char, enter_world, in, Room, with, Conn} event. Since it didn't
subsribe to it, the succeed function was never called and the user never
had a connection object added. With no connection object listening for
send events, the user never gets back any messages.