WIP: JSON logging
After getting logging to file working I'm realizing that logging to file
is next to worthless. I'm going to log everything as JSON. Users can
login as a logging connection and erlmud_conn will register that
connection as a logger (wrapped in a function to keep it opaque) with
the erlmud_event_log.

The existing web logs will probably be re-written to read JSON and
display it as it is now.