WIP: more JSON logging
My plan is to change over all the log calls throughout the code to send
JSON. I'm not exactly sure what the final keys should look like. Once I
have the web end done I'll probably catalog all the different JSON and
decide how to render it on a page. So far some common elements are
stading out:
type - what type of code event are we logging
source - the process that initiated the code event
target - the target process that the code event will affect
result - succed, fail, etc.
message - a message that was sent to a process
props - the properties of the process logging the code event

At this point I'd like to see some graphical logging based on the JSON:
maybe boxes around processes, arrows showing relations between source
and target, mouse hover for properties, etc. I really think this could
make it _way_ easier to spot what is happening in the MUD. As a bonus, I
could potentially use the JSON to animate what happened.

TODO: add timestamps to the log messages!

It was very handy having the HTML the way I have it, but it would be way
more flexible if I just logged JSON and then let whatever is reading the
logs decide how to format it.

So far I have two logging targets in mind:
1) a web page made by the test script consisting of a script that can
parse the JSON and a clump of JSON in a script. I'll do what I'm doing
now and break the web page up in to a head and tail and smush the JSON
in the middle and write that all to a file.
2) A web page with a web socket that listens for JSON log events and
adds them to the page.

(1) will be for testing where I don't have to have to fire up a page and
connect it to the MUD in order to receive events and where all the
events will be generated at once.

(2) will be for when I'm testing the MUD live and don't want to wait
until my session is done to see what's happening.