WIP: see if I can capture log fields with macros
I'd like a way to keep the message handler functions expressive while
also being able to automatically capture parts of the message that
should be logged in the JSON properties.

If I have macros with mixed case I can show which parts of the event are
captured in variables:
?Character_describe_Owner_with_Context shows that Character, Owner and
Context are bound to those variables.