Fix up a ton of errors from working on log props
This is a little premature because I'm working on using parse transforms
to specify which elements of the message should be automatically logged
and with what key; however, in order to get the parse transform to work
I needed to fix all the compile errors.

Most of this will get changed again if I get the parse transform working
as I'll get to change most of the Log = [...] to -spec macros.