Auto-log with parse transforms and macros
This is the start of a parse transform that uses function specs to
determine which elements of a message should be custom log properties
and what the name of the log property key should be.

The entire goal here is to keep messages reading like English, but also
make logging the message as JSON as easy as possible.

There's also a macro in ...attribute_look.erl that shrinks the spec down
a little so I can just specify the types for the actual message.

All message elements that have a user type in the spec will be
automatically logged as the property {SpecType, MessageElement}.

There's an example in erlmud_handler_attribute_look.erl:

-spec attempt?ATTEMPT_LOG({source(), type(), target(), atom(), context()});
             ?ATTEMPT_LOG({source(), type(), target(), atom(), context()}).
attempt({#parents{owner = Owner},
         Props,
         {Source, describe, Owner, with, Context}}) ->

The ?ATTEMPT macro translates to:
-spec attempt({#parents{},
               proplist(),
               {source(), type(), target(), atom(), context()}}) ->
          {atom(), boolean(), proplist(), proplist()};
             ({#parents{},
               proplist(),
               {source(), type(), target(), atom(), context()}}) ->
          {atom(), boolean(), proplist(), proplist()}.

The key here is the message part:
  {source(), type(), target(), atom(), context()}
These line up with the message elements:
  {Source, describe, Owner, with Context}

Since source(), type(), target() and context() are all user types,
they'll be paired up with their respective message elements:

[{source, Source},
 {type, describe},
 {target, Owner},
 {context, Context}]

The 'with' atom in the message is ignored because it has a built-in type
atom().

The log properties will then be passed back to erlmud_object to be added
to the standard log properties.