Add log properties to handlers
All event handlers now pass back customer logging properties to
erlmud_object in the attempt, succeed and fail callbacks where
appropriate. Most of the messages have common properties and all event
handlers have the "type" property which is the name of the event such as
'attack', 'die', 'drop', 'get', etc. The most common log properties are
'source' and 'target' which record which object was performing the
object at/on/to the receiving object. Another common set of properties
for room movement is 'to' and 'from'. Attack events often have 'vector'
which records what is used to do the attack. The 'look' and 'describe'
events often have a 'context' property to record what text the event
should add to: e.g. "Bob: " for describing something that belongs to the
character Bob. (Descriptions are sent back to the user piecemeal so the
user will need to see some context in case other messages arrive in
between descriptions.) Attack events might also have 'calc_hit',
'calc_damage', 'hit' and 'damage' which are hopefully obvious.

I've ditched the parse_transform which used function specs to specify
the keys for the log properties. It's just as much work as manually
created the log proplists, but _way_ harder to figure out.

I also cleaned out (hopefully) all of the log/1 functions in the
handlers since erlmud_object now does all the logging since it has most
of the log properties that are used.