Run div height handlers after body is loaded
When the log DIVs are first being setup their clientHeight property
doesn't seem to be set yet. Each "handler" span returns a handler
function (meant to refer to handling the div height, not pertaining to
the Erlmud event handler) that will be called with the height of the
handler span's div parent "clientHeight". Once the handler span has that
it can calculate it's top function. The reason I did this was because an
empty span with a border gets moved to the top of the div, which looks
goofy.

I think I could also just put a space in the handler span and it would
show up as normal. I don't know why I didn't do that.