Add process names to logged properties
If a PID in the logged properties has a name then include it in the list
of properties.

Also look in the properties of the log object as well as the 'props'
property of the log object, which is the properties of the erlmud_object
that triggered the log event. This fixed the issue where many of the log
fields that I expected weren't getting filled in.