Fix PID colors, move result span, dflt PID to __
% (modulo) has a higher precedence than >> (right shift) so I need to
wrap that in parentheses.

The result span should come close to the end.

Default PIDs that are missing to a double black underscore both for the
name and for the "ID" (PID converted to characters).

I built the reverse_int/1 function in the browser console but forgot to
remove the constant value I was using. D'oh!

add_pid/3 was using the wrong key for the type field. The variable I was
using didn't exist and the browser didn't complain even though I'm using
"use strict"; which is odd.