Map object IDs to icons
Since the logger looks up the IDs of objects in the index when logging
I could map the IDs to icons.

I could create a new index, or add the icon to the index as well, ...
and I might eventually do that. For now I'm just going to map them in
the logging web page.

I need to add the Creative Commons attributions to the page still.