Width of room span changes: adjust 'room' text
Since the width of the room span will change with the width of the room
ID I'm setting the negative margin of the 'room' title span dynamically
based on the room ID width once the page has loaded.

I can re-use the handlers to run after the line is loaded when I'm
receiving log events from a live server.